Moved intensity error to integrator
